+++ db/src/index.ts
@@ -1,19 +1,9 @@
 import * as AWS from "aws-sdk";
 import productsMock from "../../product-service/src/mocks/products.json";
-import { MAX_REQS, REGION, TableNames } from "./constants";
-import { logResult, makePutReq } from "./utils";
+import { REGION, TableNames } from "./constants";
+import { makePutReq, sendItems, sendLimitted } from "./utils";
 
 AWS.config.update({ region: REGION });
-
-const ddb = new AWS.DynamoDB();
-
-const sendItems = (reqItems: AWS.DynamoDB.BatchWriteItemRequestMap) => {
-  const params: AWS.DynamoDB.BatchWriteItemInput = {
-    RequestItems: reqItems,
-  };
-  // console.log(JSON.stringify(params, null, 2));
-  ddb.batchWriteItem(params, logResult);
-};
 
 const fill = () => {
   const productRequests = productsMock.map((p) => makePutReq(p));
@@ -40,27 +30,6 @@ const fill = () => {
   );
   if (count2) {
     sendItems(reqItems2);
-  }
-
-  function sendLimitted(
-    reqs: AWS.DynamoDB.WriteRequest[],
-    tableName: TableNames,
-    reqItems: AWS.DynamoDB.BatchWriteItemRequestMap,
-    count: number
-  ): [number, AWS.DynamoDB.BatchWriteItemRequestMap] {
-    for (const req of reqs) {
-      if (!reqItems[tableName]) {
-        reqItems[tableName] = [];
-      }
-      reqItems[tableName].push(req);
-      count++;
-      if (count === MAX_REQS) {
-        sendItems(reqItems);
-        reqItems = {};
-        count = 0;
-      }
-    }
-    return [count, reqItems];
   }
 };
