+++ db/src/utils.ts
@@ -1,4 +1,5 @@
 import * as AWS from "aws-sdk";
+import { MAX_REQS, TableNames } from "./constants";
 
 export const getValType = (val: any): string => {
   if (typeof val === "number") {
@@ -57,4 +58,34 @@ export const getFillParams = (
     },
   };
   ddb.batchWriteItem(params, logResult);
+};
+
+export const sendItems = (reqItems: AWS.DynamoDB.BatchWriteItemRequestMap) => {
+  const params: AWS.DynamoDB.BatchWriteItemInput = {
+    RequestItems: reqItems,
+  };
+  // console.log(JSON.stringify(params, null, 2));
+  const ddb = new AWS.DynamoDB();
+  ddb.batchWriteItem(params, logResult);
+};
+
+export const sendLimitted = (
+  reqs: AWS.DynamoDB.WriteRequest[],
+  tableName: TableNames,
+  reqItems: AWS.DynamoDB.BatchWriteItemRequestMap,
+  count: number
+): [number, AWS.DynamoDB.BatchWriteItemRequestMap] => {
+  for (const req of reqs) {
+    if (!reqItems[tableName]) {
+      reqItems[tableName] = [];
+    }
+    reqItems[tableName].push(req);
+    count++;
+    if (count === MAX_REQS) {
+      sendItems(reqItems);
+      reqItems = {};
+      count = 0;
+    }
+  }
+  return [count, reqItems];
 };
